Refactor data._helpers
